// frontend/src/api/generated/models/projectsIndexSortBy.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Laravel
 * OpenAPI spec version: 0.0.1
 */

export type ProjectsIndexSortBy =
  (typeof ProjectsIndexSortBy)[keyof typeof ProjectsIndexSortBy] | null

export const ProjectsIndexSortBy = {
  id: 'id',
  client_name: 'client_name',
  project_name: 'project_name',
  status: 'status',
  priority: 'priority',
  start_date: 'start_date',
  due_date: 'due_date',
  created_at: 'created_at',
} as const