// frontend/src/api/generated/models/projectsIndexSortOrder.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Laravel
 * OpenAPI spec version: 0.0.1
 */

export type ProjectsIndexSortOrder =
  (typeof ProjectsIndexSortOrder)[keyof typeof ProjectsIndexSortOrder] | null

export const ProjectsIndexSortOrder = {
  asc: 'asc',
  desc: 'desc',
  ASC: 'ASC',
  DESC: 'DESC',
} as const